refactor: major refactor and file renaming
