+++ README.md
@@ -56,7 +56,7 @@ All packages are placed in the [`packages/`](./packages) directory.
 | [`@aries-framework/rest`](https://www.npmjs.com/package/@aries-framework/rest)                             | ![@aries-framework/rest version](https://img.shields.io/npm/v/@aries-framework/rest)                             | REST endpoint wrapper for using your agent over HTTP    |
 | [`@aries-framework/react-hooks`](https://www.npmjs.com/package/@aries-framework/react-hooks)               | ![@aries-framework/react-hooks version](https://img.shields.io/npm/v/@aries-framework/react-hooks)               | React Hooks for data handling and agent interaction     |
 | [`@aries-framework/redux-store`](https://www.npmjs.com/package/@aries-framework/redux-store)               | ![@aries-framework/redux-store version](https://img.shields.io/npm/v/@aries-framework/redux-store)               | Redux Toolkit wrapper around Aries Framework JavaScript |
-| [`@aries-framework/push-notifications`](https://www.npmjs.com/package/@aries-framework/push-notifications) | ![@aries-framework/push-notifications version](https://img.shields.io/npm/v/@aries-framework/push-notifications) | Push notification plugin for Aries Framework JavaScript |
+| [`@aries-framework/push-notifications`](https://www.npmjs.com/package/@aries-framework/push-notifications) | ![@aries-framework/push-notifications version](https://img.shields.io/npm/v/@aries-framework/push-notifications) | Push notification plugin for Aries Framework JavaScript | [`@aries-framework/transport-ble`](https://www.npmjs.com/package/@aries-framework/transport-ble) | ![@aries-framework/transport-ble version](https://img.shields.io/npm/v/@aries-framework/transport-ble) | Bluetooth Low Energy transport for Aries Framework JavaScript |
 
 ## Contributing
