+++ packages/transport-ble/jest.config.ts
