+++ packages/transport-ble/package.json
@@ -25,9 +25,6 @@
     "prepublishOnly": "yarn run build",
     "test": "jest"
   },
-  "dependencies": {
-    "@animo-id/react-native-ble-didcomm": "0.3.0"
-  },
   "devDependencies": {
     "@aries-framework/core": "0.3.2",
     "@types/react-native": "0.70.8",
@@ -35,6 +32,7 @@
     "typescript": "~4.4.2"
   },
   "peerDependencies": {
+    "@animo-id/react-native-ble-didcomm": "0.3.0",
     "@aries-framework/core": "^0.3.2",
     "react-native": ">=0.60.0"
   }
