+++ packages/transport-ble/src/transports/BleOutboundTransport.ts
