+++ packages/transport-ble/src/types/index.ts
