+++ packages/transport-ble/tsconfig.json
